Fix initial Cypress test
Add a class to the GitHub login button so it can be retrieved by the test.
Tests

Added tests to the repository. The majority of the tests are unit tests run in JEST. There is a bare framework for running integration tests with Cypress. Tests are configured to run automatically using GitHub Actions.