Choose template
Allow users the choice of the template used to create a new workshop. Templates are existing repositories tagged with the topic 'ukrn-wb-template'. They should have a blank workshop_id in `./_config.yml` so that users are prompted to customise the workshop_id and other settings once their copy of the template has been created.